fix more implicit conversions

+++ src/particles/profiles/GetInitialDensity.cpp
@@ -143,7 +143,7 @@ PlasmaDensityAccessor::define_from_file (const std::string& path, std::shared_pt
 
     for (int i=0; i<3; ++i) {
         m_strides[i] = idx_perm[i] != -1 ? strides[idx_perm[i]] : 0;
-        m_bigend[i] = idx_perm[i] != -1 ? extent[idx_perm[i]] - 1 : 0;
+        m_bigend[i] = idx_perm[i] != -1 ? static_cast<int>(extent[idx_perm[i]]) - 1 : 0;
         m_pos_offset[i] = idx_perm[i] != -1 ? static_cast<amrex::Real>(
             offset[idx_perm[i]] + spacing[idx_perm[i]] * position[idx_perm[i]]) : 0;
         m_dx_inv[i] = idx_perm[i] != -1 ? static_cast<amrex::Real>(1. / spacing[idx_perm[i]]) : 0;
@@ -151,7 +151,7 @@ PlasmaDensityAccessor::define_from_file (const std::string& path, std::shared_pt
 
     if (use_mode) {
         m_strides[2] = mode_stride;
-        m_bigend[2] = mode_bigend;
+        m_bigend[2] = static_cast<int>(mode_bigend);
     }
 
     m_unitSI = static_cast<amrex::Real>(comp.unitSI());
